Create SignIn.tsx for Signing in to Application
Sign in authenticates a user by requesting their email and password.

User credentials are posted to the server, if the user exists and credentials are successful a token will be returned to the client an stored in App.tsx state to be used by other components.

Has 1 Warning:

When token is undefined, routing to packages/index produces this error after the router pushes user to /signin

Warning: Cannot update during an existing state transition (such as within `render`). Render methods should be a pure function of props and state.